requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==3.1.0         # via google-auth
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.9.0  # via google-cloud-core, google-cloud-datastore, google-cloud-storage, google-cloud-tasks
google-api-python-client==1.7.8  # via gumo-core
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2
google-cloud-core==0.29.1  # via google-cloud-datastore, google-cloud-storage
google-cloud-datastore==1.7.3  # via gumo-datastore
google-cloud-storage==1.15.0  # via gumo-core
google-cloud-tasks==1.1.0
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos[grpc]==1.5.9  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.11.4  # via google-cloud-tasks
grpcio==1.20.0            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
gumo-core==0.0.31
gumo-datastore==0.0.16
httplib2==0.12.3          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
injector==0.15.0          # via gumo-core
protobuf==3.7.1           # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pytz==2019.1              # via google-api-core
pyyaml==5.1               # via gumo-core
requests==2.21.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, protobuf
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.24.2           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via google-api-core, protobuf
